Make this file compiles on alpha even if FW_DEBUG is defined.

+++ sys/dev/ieee1394/fwohci.c
@@ -1110,9 +1110,9 @@ fwohci_buf_pktget(struct fwohci_softc *sc, struct fwohci_ctx *fc, caddr_t *pp,
 	fd = fb->fb_desc;
 #ifdef FW_DEBUG
 	if (fw_verbose)
-		printf("fwohci_buf_pktget: desc %d, off %d, req %d, res %d,"
+		printf("fwohci_buf_pktget: desc %ld, off %d, req %d, res %d,"
 		    " len %d, avail %d\n",
-		    fd - sc->sc_desc, fb->fb_off, fd->fd_reqcount,
+		    (long)(fd - sc->sc_desc), fb->fb_off, fd->fd_reqcount,
 		    fd->fd_rescount, len,
 		    fd->fd_reqcount - fd->fd_rescount - fb->fb_off);
 #endif
@@ -1355,8 +1355,9 @@ fwohci_handler_set(struct fwohci_softc *sc,
 		fwohci_ctx_init(sc, fc);
 #ifdef FW_DEBUG
 		if (fw_verbose)
-			printf("fwohci_handler_set: SYNC desc %d\n",
-			    TAILQ_FIRST(&fc->fc_buf)->fb_desc - sc->sc_desc);
+			printf("fwohci_handler_set: SYNC desc %ld\n",
+			    (long)(TAILQ_FIRST(&fc->fc_buf)->fb_desc -
+			    sc->sc_desc));
 #endif
 		OHCI_SYNC_RX_DMA_WRITE(sc, fc->fc_ctx, OHCI_SUBREG_ContextMatch,
 		    (OHCI_CTXMATCH_TAG0 << key2) | key1);
@@ -1664,7 +1665,8 @@ fwohci_at_output(struct fwohci_softc *sc, struct fwohci_ctx *fc,
 
 #ifdef FW_DEBUG
 	if (fw_verbose) {
-		printf("fwohci_at_output: desc %d", fb->fb_desc - sc->sc_desc);
+		printf("fwohci_at_output: desc %ld",
+		    (long)(fb->fb_desc - sc->sc_desc));
 		for (i = 0; i < ndesc * 4; i++)
 			printf("%s%08x", i&7?" ":"\n\t",
 			    ((u_int32_t *)fb->fb_desc)[i]);
@@ -1709,9 +1711,9 @@ fwohci_at_done(struct fwohci_softc *sc, struct fwohci_ctx *fc, int force)
 		fd = fb->fb_desc;
 #ifdef FW_DEBUG
 		if (fw_verbose) {
-			printf("fwohci_at_done: %sdesc %d (%d)",
+			printf("fwohci_at_done: %sdesc %ld (%d)",
 			    force ? "force " : "",
-			    fd - sc->sc_desc, fb->fb_nseg);
+			    (long)(fd - sc->sc_desc), fb->fb_nseg);
 			for (i = 0; i < fb->fb_nseg * 4; i++)
 				printf("%s%08x", i&7?" ":"\n    ",
 				    ((u_int32_t *)fd)[i]);
